requirements.txt
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.1.0
Pillow==10.3.0